+++ R/block.R
@@ -16,18 +16,43 @@ call_block = function(block) {
   af = opts_knit$get('eval.after'); al = opts_knit$get('aliases')
   if (!is.null(al) && !is.null(af)) af = c(af, names(al[af %in% al]))
 
-  # expand parameters defined via template
-  if (!is.null(block$params$opts.label)) {
-    block$params = merge_list(opts_template$get(block$params$opts.label), block$params)
+  params = opts_chunk$merge(block$params)
+  for (o in setdiff(names(params), af)) {
+    params[o] = list(eval_lang(params[[o]]))
+    # also update original options before being merged with opts_chunk
+    if (o %in% names(block$params)) block$params[o] = params[o]
   }
 
-  params = opts_chunk$merge(block$params)
-  opts_current$restore(params)
-  for (o in setdiff(names(params), af)) params[o] = list(eval_lang(params[[o]]))
-
   label = ref.label = params$label
-  if (!is.null(params$ref.label)) ref.label = sc_split(params$ref.label)
+  if (!is.null(params$ref.label)) {
+    ref.label = sc_split(params$ref.label)
+    # ref.label = I() implies opts.label = ref.label
+    if (inherits(params$ref.label, 'AsIs') && is.null(params$opts.label))
+      params$opts.label = ref.label
+  }
   params[["code"]] = params[["code"]] %n% unlist(knit_code$get(ref.label), use.names = FALSE)
+
+  # opts.label = TRUE means inheriting chunk options from ref.label
+  if (isTRUE(params$opts.label)) params$opts.label = ref.label
+  # expand chunk options defined via opts_template and reference chunks
+  params2 = NULL
+  for (lab in params$opts.label) {
+    # referenced chunk options (if any) override template options
+    params3 = merge_list(opts_template$get(lab), attr(knit_code$get(lab), 'chunk_opts'))
+    params2 = merge_list(params2, params3)
+  }
+  if (length(params2)) {
+    # local options override referenced options
+    params2 = merge_list(params2, block$params)
+    # then override previously merged opts_chunk options
+    params  = merge_list(params, params2)
+    # in case any options are not evaluated
+    for (o in setdiff(names(params), af)) params[o] = list(eval_lang(params[[o]]))
+  }
+
+  # save current chunk options in opts_current
+  opts_current$restore(params)
+
   if (opts_knit$get('progress')) print(block)
 
   if (!is.null(params$child)) {
@@ -253,9 +278,8 @@ eng_r = function(options) {
   # number of plots in this chunk
   if (is.null(options$fig.num))
     options$fig.num = if (length(res)) sum(sapply(res, function(x) {
-      if (evaluate::is.recordedplot(x)) return(1)
       if (inherits(x, 'knit_image_paths')) return(length(x))
-      if (inherits(x, 'html_screenshot')) return(1)
+      if (is_plot_output(x)) return(1)
       0
     })) else 0L
 
@@ -270,7 +294,7 @@ eng_r = function(options) {
     opts_knit$delete('plot_files')
   }, add = TRUE)  # restore plot number
 
-  output = unlist(wrap(res, options)) # wrap all results together
+  output = unlist(sew(res, options)) # wrap all results together
   res.after = run_hooks(before = FALSE, options, env) # run 'after' hooks
 
   output = paste(c(res.before, output, res.after), collapse = '')  # insert hook results
@@ -364,7 +388,7 @@ chunk_device = function(options, record = TRUE, tmp = tempfile()) {
 
 # fall back to a usable device (e.g., during R CMD check)
 fallback_dev = function(dev) {
-  if (length(dev) != 1 || !getOption('knitr.device.fallback', xfun::is_R_CMD_check()))
+  if (length(dev) != 1 || !getOption('knitr.device.fallback', is_R_CMD_check()))
     return(dev)
   choices = list(
     svg = c('png', 'jpeg', 'bmp'), cairo_pdf = c('pdf'), cairo_ps = c('postscript'),
@@ -397,7 +421,8 @@ find_recordedplot = function(x) {
 }
 
 is_plot_output = function(x) {
-  evaluate::is.recordedplot(x) || inherits(x, 'knit_image_paths')
+  evaluate::is.recordedplot(x) ||
+    inherits(x, c('knit_image_paths', 'html_screenshot', 'knit_other_plot'))
 }
 
 # move plots before source code
@@ -501,7 +526,7 @@ inline_exec = function(
   loc = block$location
   for (i in 1:n) {
     res = hook_eval(code[i], envir)
-    if (inherits(res, 'knit_asis')) res = wrap(res, inline = TRUE)
+    if (inherits(res, 'knit_asis')) res = sew(res, inline = TRUE)
     d = nchar(input)
     # replace with evaluated results
     stringr::str_sub(input, loc[i, 1], loc[i, 2]) = if (length(res)) {